Add HDF5 based dataset option (#227)
