+++ .github/workflows/tests_develop.yml
@@ -32,6 +32,7 @@ jobs:
         python -m pip install --upgrade pip
         pip install setuptools wheel
         pip install torch==${TORCH} -f https://download.pytorch.org/whl/cpu/torch_stable.html
+        pip install h5py
         pip install --upgrade-strategy only-if-needed .
     - name: Install pytest
       run: |
